ResNet18/requirements.txt
# ===============================
#  Core Deep Learning Libraries
# ===============================
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0

# ===============================
#  Data Handling & Utilities
# ===============================
numpy>=1.24.0
pandas>=2.2.0
tqdm>=4.66.0

# ===============================
#  Image Processing
# ===============================
Pillow>=10.0.0

# ===============================
#  Visualization (optional)
# ===============================
matplotlib>=3.8.0

# ===============================
#  Logging (optional, for TensorBoard)
# ===============================
tensorboard>=2.14.0
